updated setup, moved foundry int tests and updated merge test
