+++ .github/workflows/python-merge-tests.yml
@@ -70,6 +70,7 @@ jobs:
   python-tests-azure:
     name: Python Tests - Azure
     runs-on: ${{ matrix.os }}
+    environment: "integration"
     strategy:
       fail-fast: true
       matrix:
@@ -97,6 +98,13 @@ jobs:
       - name: Install the project
         run: |
           uv sync --all-packages --all-extras --dev -U --prerelease=if-necessary-or-explicit
+      - name: Azure CLI Login
+        if: github.event_name != 'pull_request'
+        uses: azure/login@v2
+        with:
+          client-id: ${{ secrets.AZURE_CLIENT_ID }}
+          tenant-id: ${{ secrets.AZURE_TENANT_ID }}
+          subscription-id: ${{ secrets.AZURE_SUBSCRIPTION_ID }}
       - name: Test with pytest
         run: uv run poe --directory ./packages/${{ env.PACKAGE_NAME }} test --junitxml=coverage.xml
         working-directory: ./python
@@ -122,6 +130,7 @@ jobs:
   python-tests-foundry:
     name: Python Tests - Foundry
     runs-on: ${{ matrix.os }}
+    environment: "integration"
     strategy:
       fail-fast: true
       matrix:
@@ -129,8 +138,8 @@ jobs:
         os: [ubuntu-latest]
     env:
       UV_PYTHON: ${{ matrix.python-version }}
-      AZURE_OPENAI_CHAT_DEPLOYMENT_NAME: ${{ vars.AZUREAI__DEPLOYMENTNAME }}
-      AZURE_OPENAI_ENDPOINT: ${{ secrets.AZUREAI__ENDPOINT }}
+      FOUNDRY_PROJECT_ENDPOINT: ${{ vars.FOUNDRY_PROJECT_ENDPOINT }}
+      FOUNDRY_MODEL_DEPLOYMENT_NAME: ${{ vars.FOUNDRY_MODEL_DEPLOYMENT_NAME }}
       PACKAGE_NAME: "foundry"
     permissions:
       contents: write
@@ -149,6 +158,13 @@ jobs:
       - name: Install the project
         run: |
           uv sync --all-packages --all-extras --dev -U --prerelease=if-necessary-or-explicit
+      - name: Azure CLI Login
+        if: github.event_name != 'pull_request'
+        uses: azure/login@v2
+        with:
+          client-id: ${{ secrets.AZURE_CLIENT_ID }}
+          tenant-id: ${{ secrets.AZURE_TENANT_ID }}
+          subscription-id: ${{ secrets.AZURE_SUBSCRIPTION_ID }}
       - name: Test with pytest
         run: uv run poe --directory ./packages/${{ env.PACKAGE_NAME }} test --junitxml=coverage.xml
         working-directory: ./python
