+++ .github/workflows/python-tests.yml
@@ -1,4 +1,4 @@
-name: Python - Tests - No Env
+name: Python - Tests
 
 on:
   pull_request:
@@ -11,7 +11,7 @@ env:
 
 jobs:
   python-tests:
-    name: Python Tests - No Env
+    name: Python Tests
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: true
