+++ python/packages/foundry/tests/conftest.py
@@ -20,6 +20,7 @@ def override_env_param_dict(request: Any) -> dict[str, str]:
 @fixture()
 def foundry_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):  # type: ignore
     """Fixture to set environment variables for FoundrySettings."""
+
     if exclude_list is None:
         exclude_list = []
 
@@ -35,10 +36,10 @@ def foundry_unit_test_env(monkeypatch, exclude_list, override_env_param_dict):  
     env_vars.update(override_env_param_dict)  # type: ignore
 
     for key, value in env_vars.items():
-        if key not in exclude_list:
-            monkeypatch.setenv(key, value)  # type: ignore
-        else:
+        if key in exclude_list:
             monkeypatch.delenv(key, raising=False)  # type: ignore
+            continue
+        monkeypatch.setenv(key, value)  # type: ignore
 
     return env_vars
