--- python/packages/main/tests/test_version.py
@@ -1,8 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-
-from agent_framework import __version__
-
-
-def test_version():
-    assert __version__ is not None
